Extract extendTableData to a utils file
